+++ src/hazardous/ecc/x25519.rs
@@ -107,7 +107,7 @@ pub(crate) struct FieldElement(fiat_25519_tight_field_element);
 
 impl core::fmt::Debug for FieldElement {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        write!(f, "FieldElement({:?})", &self.0.0)
+        write!(f, "FieldElement({:?})", self.0.0)
     }
 }